Refactor the request utilities
