--- app/requests/request/movement_type.rb
@@ -1,6 +1,0 @@
-module Request
-  class MovementType < Request::Base
-    @entity = "movement_types"
-
-  end
-end
