--- app/requests/request/speciality.rb
@@ -1,7 +1,0 @@
-module Request
-  class Speciality < Request::Base
-    @entity = "specialities"
-
-    
-  end
-end
