+++ config/application.rb
@@ -37,5 +37,6 @@ module CentersApi
     #Requests router
     config.paths.add File.join('app', 'requests'), glob: File.join('**', '*.rb')
     config.autoload_paths += Dir[Rails.root.join('app', 'requests', '*')]
+    config.autoload_paths += %W(#{config.root}/lib)
   end
 end
